fix: move tree to internal
