+++ internal/tree/children.go
