+++ internal/tree/enumerator.go
