+++ internal/tree/renderer.go
