+++ internal/tree/tree_test.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/aymanbagabas/go-udiff"
 	"github.com/charmbracelet/lipgloss"
+	"github.com/charmbracelet/lipgloss/internal/tree"
 	"github.com/charmbracelet/lipgloss/list"
 	"github.com/charmbracelet/lipgloss/table"
-	"github.com/charmbracelet/lipgloss/tree"
 )
 
 func TestTree(t *testing.T) {
