+++ list/list.go
@@ -24,7 +24,7 @@ package list
 
 import (
 	"github.com/charmbracelet/lipgloss"
-	"github.com/charmbracelet/lipgloss/tree"
+	"github.com/charmbracelet/lipgloss/internal/tree"
 )
 
 // List represents a list of items that can be displayed. Lists can contain
